Wand of anti-magic now charges up and deals damage to non-magic foes
Mobs now have a MAGIC property (or don't). Those who are magic are
given an anti-magic buff, and the wand charges up a bit. When the wand
zaps a mob who is not magic, or who has an anti-magic buff, they are
dealt damage proportional to the stored magic energy.

The current energy level of the wand is not yet observable.