Adjust carpet charges, make speed variable
Previously, the carpet's maximum charge stopped increasing at level
7, making further upgrades not useful. This was ok when the carpet
upgraded itself, but now that you have to upgrade it through alchemy,
it's important to get something in return.

Previously, the carpet's speed was 3, the same as with a Potion of
Haste. Now it starts at 2, and is upgradable to 4.

Here's the full schedule of upgrades. Range is how far you can go
on a full charge and is just Charges x Speed. Upgrade Cost is the
energy required in an alchemy pot, in addition to Potions of Haste
and Levitation.

                              Upgrade
Level  Charges  Speed  Range   Cost
  0      3       2.0      6      4
  2      4       2.4     10      5
  4      5       2.8     14      6
  6      6       3.2     19      7
  8      7       3.6     25      8
 10      8       4.0     32