Use hero's flying carpet in alchemy recipe guide
If the hero has a flying carpet, the QuickRecipe alchemy guide uses
their carpet as the recipe input, an appropriately upgraded one as the
output, and shows the corresponding cost.

Otherwise we use unidentified carpets for input and output, which
don't show the level, and the recipe cost is for a level zero upgrade.

This is unclear, in that doesn't show any upgrade effect (because the
output level is not shown), and it doesn't illustrate the variable
cost, but it's clean, simple, not misleading, and gets the basic idea
across.

Maybe it would be better to show an identified level zero carpet as
input, and an identified upgraded carpet in the output as that would
make the point of it all clearer. Let's try this and get some feedback.