Fix flying carpet alchemy duplication bug
Alchemy was not designed to work with equippable items. This fixes a
bug where the carpet was not being unequipped when the alchemy pot was
populated using the QuickRecipe from the guide.

(It was instead looking to see if the duplicate copy in the recipe
inputs was equipped, which it never is.)

Eventually it would be nice to re-equip items when removing them from
the pot, but for now the goal is just to get this to work.