Disable custom notes in the forgetful challenge
The forgetful challenge is not very challenging (or fun, or interesting)
if you make notes on each inventory item. We can't keep people from
making notes outside the game, but having the game track everything for
you is too easy and runs counter to the challenge, so .. no notes!

The regular journal entries still work, and the hero doesn't forget
level maps... yet.