Disable the journal in the forgetful challenge
The forgetful challenge is all about having to remember things. A
previous commit (unreleased) disabled custom notes on inventory items,
as that would make it too easy to track potions and scrolls. This change
goes further and also removes the log of landmarks on various levels.

We can't keep players from making notes outside the game, but having
the game track everything for you is too easy and runs counter to the
challenge, so no more journal.